# Commerce-Hub/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeRoute B2B marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products visible to the caller
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: manufacturerId
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (manufacturer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/{productId}/platform-price:
    patch:
      operationId: setProductPlatformPrice
      tags: [products]
      summary: Admin — override the shopkeeper-facing price for a product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPlatformPriceInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List shipments relevant to the caller
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"
  /shipments/{shipmentId}:
    get:
      operationId: getShipment
      tags: [shipments]
      parameters:
        - in: path
          name: shipmentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /shipments/{shipmentId}/status:
    post:
      operationId: updateShipmentStatus
      tags: [shipments]
      summary: Transporter updates shipment status
      parameters:
        - in: path
          name: shipmentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShipmentStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /shipments/{shipmentId}/location:
    post:
      operationId: updateShipmentLocation
      tags: [shipments]
      summary: Transporter updates current GPS location of the shipment
      parameters:
        - in: path
          name: shipmentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationInput"
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Product:
      type: object
      required:
        - id
        - name
        - description
        - category
        - unit
        - pricePerUnit
        - minOrderQty
        - stockQty
        - manufacturerId
        - manufacturerName
        - createdAt
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        unit:
          type: string
        pricePerUnit:
          type: number
        platformPrice:
          type: number
          nullable: true
        minOrderQty:
          type: integer
        stockQty:
          type: integer
        imageUrl:
          type: string
          nullable: true
        manufacturerId:
          type: string
        manufacturerName:
          type: string
        createdAt:
          type: string
          format: date-time
    SetPlatformPriceInput:
      type: object
      properties:
        platformPrice:
          type: number
          nullable: true
    CreateProductInput:
      type: object
      required: [name, description, category, unit, pricePerUnit, minOrderQty, stockQty]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        unit:
          type: string
        pricePerUnit:
          type: number
        minOrderQty:
          type: integer
        stockQty:
          type: integer
        imageUrl:
          type: string
          nullable: true
    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        unit:
          type: string
        pricePerUnit:
          type: number
        minOrderQty:
          type: integer
        stockQty:
          type: integer
        imageUrl:
          type: string
          nullable: true
...
    ShipmentStatus:
      type: string
      enum: [assigned, picked_up, in_transit, delivered]
    Shipment:
      type: object
      required:
        - id
        - trackingNumber
        - status
        - orderId
        - orderNumber
        - transporterId
        - transporterName
        - originAddress
        - destinationAddress
        - createdAt
      properties:
        id:
          type: string
        trackingNumber:
          type: string
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        orderId:
          type: string
        orderNumber:
          type: string
        transporterId:
          type: string
        transporterName:
          type: string
        transporterCompany:
          type: string
          nullable: true
        shopkeeperName:
          type: string
          nullable: true
        shopkeeperCompany:
          type: string
          nullable: true
        shopkeeperPhone:
          type: string
          nullable: true
        manufacturerName:
          type: string
          nullable: true
        manufacturerCompany:
          type: string
          nullable: true
        manufacturerPhone:
          type: string
          nullable: true
        originAddress:
          type: string
        destinationAddress:
          type: string
        pickupDate:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        originLat:
          type: number
          nullable: true
        originLng:
          type: number
          nullable: true
        destLat:
          type: number
          nullable: true
        destLng:
          type: number
          nullable: true
        currentLat:
          type: number
          nullable: true
        currentLng:
          type: number
          nullable: true
        lastLocationAt:
          type: string
          format: date-time
          nullable: true
        pickupPin:
          type: string
          nullable: true
        deliveryPin:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    UpdateShipmentStatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        pin:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    UpdateLocationInput:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
...